hacking of the gc
I have removed all the garbage collection,
the idea is that we will not need to free any memory for now until we
can figure things out. the GC was crashing gdb.

I want to keep all the memory and later have an account for them what
is in there and how they are used.